EPITA_DataViz_Most-Dangerous-Country-in-EU
We use crimes statistics from europea.eu website to explain or show the most dangerouns countries in European Union. We used 4 datasets, one for the Assaults, another for the number of intentional Homicide.another dataset of the number of car thefts and one last with the number of the number of robbery.